updated the assignment api and splited the streakPage component into multiple reusable subcomponents
