+++ src/services/api/ApiHandler.ts
@@ -1,6 +1,6 @@
 import {
   AssignmentCartData,
-  AssignmentBatchGroupRow,
+  AssignmentDateRangeData,
   CampaignAudienceOptions,
   CampaignAudiencePayload,
   CampaignAudienceSummary,
@@ -1233,26 +1233,13 @@ export class ApiHandler implements ServiceApi {
       endDate,
     );
   }
-  async getAssignmentsForClassAndSchoolByDateRange(
+  async getAssignmentDateRangeDataForClassAndSchool(
     classId: string,
     schoolId: string,
     startDate: string,
     endDate: string,
-  ): Promise<TableTypes<'assignment'>[]> {
-    return this.s.getAssignmentsForClassAndSchoolByDateRange(
-      classId,
-      schoolId,
-      startDate,
-      endDate,
-    );
-  }
-  async getAssignmentBatchGroupsForClassAndSchoolByDateRange(
-    classId: string,
-    schoolId: string,
-    startDate: string,
-    endDate: string,
-  ): Promise<AssignmentBatchGroupRow[]> {
-    return this.s.getAssignmentBatchGroupsForClassAndSchoolByDateRange(
+  ): Promise<AssignmentDateRangeData> {
+    return this.s.getAssignmentDateRangeDataForClassAndSchool(
       classId,
       schoolId,
       startDate,
@@ -1268,13 +1255,13 @@ export class ApiHandler implements ServiceApi {
     return this.s.getCoinAndStreakCount(userId, classId, schoolId);
   }
 
-  async putCoins(
+  async updateCoins(
     userId: string,
     schoolId: string,
     classId: string,
     coins: number,
   ): Promise<TableTypes<TABLES.UserAchivements>> {
-    return this.s.putCoins(userId, schoolId, classId, coins);
+    return this.s.updateCoins(userId, schoolId, classId, coins);
   }
 
   getTeacherJoinedDate(
