+++ src/teachers-module/components/homePage/assignment/CreateSelectedAssignment.test.tsx
@@ -54,7 +54,7 @@ const mockApi = {
   getCoursesForClassStudent: jest.fn(),
   getUserAssignmentCart: jest.fn(),
   createOrUpdateAssignmentCart: jest.fn(),
-  getAssignmentBatchGroupsForClassAndSchoolByDateRange: jest.fn(),
+  getAssignmentDateRangeDataForClassAndSchool: jest.fn(),
   getChapterByLesson: jest.fn(),
   getCourse: jest.fn(),
   getChapterById: jest.fn(),
@@ -160,9 +160,10 @@ describe('CreateSelectedAssignment (QR flow)', () => {
       lessons: JSON.stringify({}),
     });
     mockApi.createOrUpdateAssignmentCart.mockResolvedValue(true);
-    mockApi.getAssignmentBatchGroupsForClassAndSchoolByDateRange.mockResolvedValue(
-      [],
-    );
+    mockApi.getAssignmentDateRangeDataForClassAndSchool.mockResolvedValue({
+      assignments: [],
+      batchGroups: [],
+    });
     mockApi.getChapterByLesson.mockResolvedValue('chapter-1');
     mockApi.getCourse.mockResolvedValue({ name: 'Math' });
     mockApi.getChapterById.mockResolvedValue({ name: 'Chapter 1' });
