+++ src/teachers-module/components/streakComponent/streakPage.css
@@ -29,7 +29,7 @@
   background: #fff;
   border: 2px solid #ff6b00;
   border-radius: 14px;
-  width: 17%;
+  width: 20%;
   min-height: 50px;
   margin-left: auto;
   padding: 0 10px;
